+++ level.h
@@ -18,14 +18,14 @@ class Level : public QObject{
     Q_PROPERTY(int levelId READ levelId WRITE setLevelId NOTIFY levelIdChanged)
     Q_PROPERTY(int rows READ rows WRITE setRows NOTIFY rowsChanged)
     Q_PROPERTY(int columns READ columns WRITE setColumns NOTIFY columnsChanged)
-    Q_PROPERTY(float difficulty READ difficulty WRITE setDifficulty NOTIFY difficultyChanged)
+    Q_PROPERTY(int difficulty READ difficulty WRITE setDifficulty NOTIFY difficultyChanged)
     Q_PROPERTY(QVariantList hints READ hints WRITE setHints NOTIFY hintsChanged)
 
 private:
     int c_levelId;
     int c_rows;
     int c_columns;
-    float c_difficulty;
+    int c_difficulty;
     QVariantList c_hints;
 
 public:
@@ -40,8 +40,8 @@ public:
     void setRows(int);
     int columns() const;
     void setColumns(int);
-    float difficulty() const;
-    void setDifficulty(float);
+    int difficulty() const;
+    void setDifficulty(int);
     QVariantList hints() const;
     void setHints(QVariantList &);
